+++ src/app/loadout-builder/utils.ts
@@ -2,17 +2,35 @@ import { DimItem } from 'app/inventory/item-types';
 import _ from 'lodash';
 import { ProcessItem } from './process-worker/types';
 
-/** Gets the stat tier from a stat value. */
+/** Gets the effective stat tier from a stat value, clamping between 0-10 */
 export function statTier(stat: number) {
   return _.clamp(Math.floor(stat / 10), 0, 10);
 }
 
 /**
- * Gets the stat tier plus a .5 if stat % 10 >= 5.
+ * Gets the stat tier plus a .5, without clamping even when exceeding T10 or going below T0.
  * To be used for display purposed only.
  */
 export function statTierWithHalf(stat: number) {
-  return `${_.clamp(Math.floor(stat / 10), 0, 10)}${stat % 10 >= 5 ? '.5' : ''}`;
+  const base_tier = Math.floor(stat / 10);
+  // ensure (-3 mod 10) results in a remainder of 7 and a stat of -0.5
+  const point_five = remEuclid(stat, 10) >= 5;
+  const sign = base_tier < 0 ? '-' : '';
+  return `${sign}${base_tier < 0 && point_five ? Math.abs(base_tier + 1) : Math.abs(base_tier)}${
+    point_five ? '.5' : ''
+  }`;
+}
+
+/**
+ * Calculates the remainder of euclidean division `dividend / divisor`,
+ * i.e. returns `rem` such that `dividend = divisor * n + rem` and
+ * `0 <= rem < |divisor|`.
+ * Remainder is always non-negative, behavior differs from `%` for
+ * negative dividends. Find comparisons at
+ * https://en.wikipedia.org/wiki/Modulo_operation#Variants_of_the_definition
+ */
+export function remEuclid(dividend: number, divisor: number) {
+  return ((dividend % divisor) + divisor) % divisor;
 }
 
 /**