+++ src/app/loadout/mod-assignment-utils.ts
@@ -12,6 +12,7 @@ import {
   modTypeTagByPlugCategoryHash,
 } from 'app/search/specialty-modslots';
 import { compareBy } from 'app/utils/comparators';
+import { emptyArray } from 'app/utils/empty';
 import { getModTypeTagByPlugCategoryHash, getSpecialtySocketMetadatas } from 'app/utils/item-utils';
 import { warnLog } from 'app/utils/log';
 import { getSocketByIndex, getSocketsByIndexes } from 'app/utils/socket-utils';
@@ -303,7 +304,7 @@ export function fitMostMods(
 
   const unassignedMods: PluggableInventoryItemDefinition[] = [];
   for (const item of items) {
-    // accumulate all unassigneds
+    // accumulate all unassigned mods
     for (const collection of [
       bucketIndependentAssignments[item.id],
       bucketSpecificAssignments[item.id],
@@ -357,11 +358,7 @@ export function pickPlugPositions(
   const existingModSockets = getSocketsByIndexes(item.sockets!, armorModIndexes || []).sort(
     // We are sorting so that we can assign mods to the socket with the least number of possible options
     // first. This helps with artificer mods as the socket is a subset of the other mod sockets on the item
-    compareBy((socket) =>
-      socket.socketDefinition.reusablePlugSetHash
-        ? defs.PlugSet.get(socket.socketDefinition.reusablePlugSetHash).reusablePlugItems.length
-        : 999
-    )
+    compareBy((socket) => (socket.plugSet ? socket.plugSet.plugs.length : 999))
   );
 
   for (const modToInsert of modsToInsert) {
@@ -374,9 +371,7 @@ export function pickPlugPositions(
     // TO-DO: this is naive and is going to be misleading for armor
     if (destinationSocketIndex === -1) {
       destinationSocketIndex = existingModSockets.findIndex((socket) =>
-        defs.PlugSet.get(socket.socketDefinition.reusablePlugSetHash!).reusablePlugItems.some(
-          (plugItem) => plugItem.plugItemHash === modToInsert.hash
-        )
+        socket.plugSet?.plugs.some((dimPlug) => dimPlug.plugDef.hash === modToInsert.hash)
       );
     }
 
@@ -445,6 +440,10 @@ export function createPluggingStrategy(
   // stuff we MAY apply, if we need more energy freed up
   const optionalRegains: PluggingAction[] = [];
 
+  if (!item.energy) {
+    return emptyArray();
+  }
+
   for (const assignment of assignments) {
     const destinationSocket = getSocketByIndex(item.sockets!, assignment.socketIndex)!;
     const existingModCost = destinationSocket.plugged?.plugDef.plug.energyCost?.energyCost || 0;
@@ -473,8 +472,8 @@ export function createPluggingStrategy(
 
   const operationSet: PluggingAction[] = [];
 
-  const itemTotalEnergy = item.energy!.energyCapacity;
-  let itemCurrentUsedEnergy = item.energy!.energyUsed;
+  const itemTotalEnergy = item.energy.energyCapacity;
+  let itemCurrentUsedEnergy = item.energy.energyUsed;
 
   // apply all required regains first
   for (const regainOperation of requiredRegains) {